Add missing file headers
This commit adds missing file headers and moves the DoubleValueTests
class into its own file.
