+++ test/DocumentFormat.OpenXml.Tests/SimpleTypes/BooleanValueTests.cs
@@ -1,4 +1,7 @@
-﻿namespace DocumentFormat.OpenXml.Tests.SimpleTypes
+﻿// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+namespace DocumentFormat.OpenXml.Tests.SimpleTypes
 {
     public class BooleanValueTests : OpenXmlComparableValueTypeTests<bool>
     {
