+++ test/DocumentFormat.OpenXml.Tests/SimpleTypes/DecimalValueTests.cs
@@ -1,4 +1,7 @@
-﻿namespace DocumentFormat.OpenXml.Tests.SimpleTypes
+﻿// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+namespace DocumentFormat.OpenXml.Tests.SimpleTypes
 {
     public class DecimalValueTests : OpenXmlComparableValueTypeTests<decimal>
     {
@@ -7,16 +10,6 @@
             SmallValue1 = new DecimalValue(10m);
             SmallValue2 = new DecimalValue(10m);
             LargeValue = new DecimalValue(20m);
-        }
-    }
-
-    public class DoubleValueTests : OpenXmlComparableValueTypeTests<double>
-    {
-        public DoubleValueTests()
-        {
-            SmallValue1 = new DoubleValue(10.0);
-            SmallValue2 = new DoubleValue(10.0);
-            LargeValue = new DoubleValue(20.0);
         }
     }
 }
