+++ test/DocumentFormat.OpenXml.Tests/SimpleTypes/OpenXmlComparableReferenceTypeTests.cs
@@ -1,4 +1,7 @@
-﻿using System;
+﻿// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using System;
 using Xunit;
 
 namespace DocumentFormat.OpenXml.Tests.SimpleTypes
